Extract time manipulation functions to mixin
I want to re-use them from outside of controller.
