+++ app/controllers/application_controller.rb
@@ -27,6 +27,7 @@ class ApplicationController < ActionController::Base
   include ActionView::Helpers::DateHelper
   include ApplicationHelper
   include JsHelper
+  include Mixins::TimeHelper
   helper ToolbarHelper
   helper JsHelper
   helper QuadiconHelper
@@ -696,29 +697,6 @@ class ApplicationController < ActionController::Base
       options[:time_profile_days] = nil
     end
     options[:tz] = options[:time_profile_tz]
-  end
-
-  # Create a time in a timezone
-  def create_time_in_tz(datetime, tz = nil)               # tz = nil means use user's session timzone
-    if tz && (Time.zone.nil? || tz != Time.zone.name)  # If tz passed in and not default tz
-      saved_tz = Time.zone
-      Time.zone = tz                                      # Temporarily convert to new tz and create the time object
-      t = Time.zone.parse(datetime)                       # Create the time object
-      Time.zone = saved_tz                                # Restore original default
-    else                                                  # tz not passed in or matches current tz
-      t = Time.zone.parse(datetime)                       # Create the time object
-    end
-    t
-  end
-
-  # Create a time in a timezone, return in UTC
-  def create_time_in_utc(datetime, tz = nil)                        # tz = nil means use user's session timzone
-    create_time_in_tz(datetime, tz).in_time_zone("Etc/UTC")  # Return the time in UTC
-  end
-
-  # convert time from utc to server timezone
-  def convert_time_from_utc(datetime)
-    datetime.in_time_zone(server_timezone)
   end
 
   # if authenticating or past login screen
